better audio freq accuracy and allow application-controlled logs (#33)
* log: new feature to disable panic handler

In some situations, the user of the library may want to use its
own panic handler. This is not possible currently, because panic_halt
is used by default when no log feature is enabled.

Add a new "log-none" feature that disables both logs and panic handler.

Signed-off-by: Olivier Matz <zer0@droids-corp.org>

* system: increase audio frequency accuracy

Currently, the effective audio frequency is:
  pll3_p_ck / 256 = 12329003 / 256 = 48160.168 Hz

This is not very accurate (0.33 % error). To enhance this, use a
fractional ratio for PLL3 configuration. After this change, the audio
frequency is:
  pll3_p_ck / 256 = 12287998 / 256 = 47999.992 Khz

The error is now 0.0000162 %, comparable to the precision of the
crystal.

Signed-off-by: Olivier Matz <zer0@droids-corp.org>

---------

Signed-off-by: Olivier Matz <zer0@droids-corp.org>
Co-authored-by: Olivier Matz <zer0@droids-corp.org>