system: enhance precision of PLL configuration
In commit e7f515e60ae1 ("better audio freq accuracy and allow
application-controlled logs (#33)"), the PLL configuration was updated
to provide a better precision, but this was incorrect.

The master clock is supposed to be 256 times the frequency sampling,
but a factor of 257 was used to ensure the frequency is above. Using
PllConfigStrategy::Fractional was then useless.

Change the factor to 256 and use PllConfigStrategy::FractionalNotLess to
keep the constraint of generating a clock higher than the target. From
init_clocks() log, the PLL3_P clock is 12288001, which is 0.000008 %
above the right frequency (12288000).

Link: https://www.st.com/content/ccc/resource/training/technical/product_training/group0/81/96/f3/1a/55/d9/40/52/STM32H7-System-Reset_and_clock_control_RCC/files/STM32H7-System-Reset_and_clock_control_RCC.pdf/_jcr_content/translations/en.STM32H7-System-Reset_and_clock_control_RCC.pdf
Link: https://www.st.com/content/ccc/resource/training/technical/product_training/group0/d3/c0/b0/0e/fe/eb/40/a9/STM32H7-Peripheral-Serial-Audio-Interface_SAI/files/STM32H7-Peripheral-Serial-Audio-Interface_SAI.pdf/_jcr_content/translations/en.STM32H7-Peripheral-Serial-Audio-Interface_SAI.pdf
Fixes: e7f515e60ae1 ("better audio freq accuracy and allow application-controlled logs (#33)")

Signed-off-by: Olivier Matz <zer0@droids-corp.org>